Fixed update query

+++ Stock/Stock/Products.cs
@@ -27,38 +27,42 @@ namespace Stock
         private void button2_Click(object sender, EventArgs e)
         {
             //TO-DO: CheckLogin username & Password
-            SqlConnection con = new SqlConnection(@"Data Source=PRENOSNIK\SQLEXPRESS;Initial Catalog=NiceLabel;Integrated Security=True");
-            //Insert Logic
-            con.Open();
-            bool status = false;
-            if (comboBox1.SelectedIndex == 0)
-            {
-                status = true;
-            }
-            else
-            {
-                status = false;
-            }
-            var sqlQuery = "";
-            if (IfProductsExists(con,textBox1.Text))
-            {
-                sqlQuery = @"UPDATE[Products] SET [ProductName] = '" + textBox2.Text + "' ,[ProductStatus] = '" + status + "' WHERE [ProductCode] = '" + textBox1.Text + "'";
-            }
-            else
-            {
-                sqlQuery= "@INSERT INTO[dbo].[Products]([ProductCode],[ProductName],[ProductStatus])VALUES('" + textBox1.Text + "','" + textBox2.Text + "','" + status + "')";
-            }
+            
+                SqlConnection con = Connection.getConnection();
+                //Insert Logic
+                con.Open();
+                bool status = false;
+                if (comboBox1.SelectedIndex == 0)
+                {
+                    status = true;
+                }
+                else
+                {
+                    status = false;
+                }
+                var sqlQuery = "";
+                if (IfProductsExists(con, textBox1.Text))
+                {
+                    sqlQuery = @"UPDATE[Products] SET [ProductName] = '" + textBox2.Text + "' ,[ProductStatus] = '" + status + "' WHERE [ProductCode] = '" + textBox1.Text + "'";
+                }
+                else
+                {
+                    sqlQuery = @"INSERT INTO[dbo].[Products]([ProductCode],[ProductName],[ProductStatus])VALUES('" + textBox1.Text + "','" + textBox2.Text + "','" + status + "')";
+                }
 
-            SqlCommand cmd = new SqlCommand(sqlQuery,con);
-            cmd.ExecuteNonQuery();
-            con.Close();
-            //Reading Data
-            LoadData();
+                SqlCommand cmd = new SqlCommand(sqlQuery, con);
+                cmd.ExecuteNonQuery();
+                con.Close();
+                //Reading Data
+                LoadData();
+                ResetRecords();
+            
+            
         }
         private bool IfProductsExists(SqlConnection con, string productCode)
         {
 
-            SqlDataAdapter sda = new SqlDataAdapter("Select 1 from [Products] WHERE[ProductCode]='"+productCode+"'", con);
+            SqlDataAdapter sda = new SqlDataAdapter("Select 1 from [Products] WHERE[ProductCode]='" + productCode + "'", con);
             DataTable dt = new DataTable();
             sda.Fill(dt);
             if (dt.Rows.Count > 0)
@@ -68,7 +72,7 @@ namespace Stock
         }
         public void LoadData()
         {
-            SqlConnection con = new SqlConnection(@"Data Source=PRENOSNIK\SQLEXPRESS;Initial Catalog=NiceLabel;Integrated Security=True");
+            SqlConnection con = Connection.getConnection();
             SqlDataAdapter sda = new SqlDataAdapter("Select * from [dbo].[Products]", con);
             DataTable dt = new DataTable();
             sda.Fill(dt);
@@ -93,6 +97,8 @@ namespace Stock
         private void dataGridView1_MouseDoubleClick(object sender, MouseEventArgs e)
         {
             //Select all data from Datagridview back to textbox
+            button2.Text = "Update";
+
             textBox1.Text = dataGridView1.SelectedRows[0].Cells[0].Value.ToString();
             textBox2.Text = dataGridView1.SelectedRows[0].Cells[1].Value.ToString();
             if (dataGridView1.SelectedRows[0].Cells[2].Value.ToString() == "Active")
@@ -103,31 +109,57 @@ namespace Stock
             {
                 comboBox1.SelectedIndex = 1;
             }
-            
+
         }
 
         private void button1_Click(object sender, EventArgs e)
         {
-            SqlConnection con = new SqlConnection(@"Data Source=PRENOSNIK\SQLEXPRESS;Initial Catalog=NiceLabel;Integrated Security=True");
-            var sqlQuery = "";
-
-            if (IfProductsExists(con, textBox1.Text))
+            if (Validation())
             {
-                con.Open();
-                sqlQuery = @"DELETE FROM[Products] WHERE [ProductCode] = '" + textBox1.Text + "'";
-                SqlCommand cmd = new SqlCommand(sqlQuery, con);
-                cmd.ExecuteNonQuery();
-                con.Close();
-            }
-            else
-            {
-                MessageBox.Show("Record Not Exists...!");
-            }
+                SqlConnection con = Connection.getConnection();
+                var sqlQuery = "";
+                if (IfProductsExists(con, textBox1.Text))
+                {
+                    con.Open();
+                    sqlQuery = @"DELETE FROM[Products] WHERE [ProductCode] = '" + textBox1.Text + "'";
+                    SqlCommand cmd = new SqlCommand(sqlQuery, con);
+                    cmd.ExecuteNonQuery();
+                    con.Close();
+                }
+                else
+                {
+                    MessageBox.Show("Record Not Exists...!");
+                }
 
-            
-
+                LoadData();
+            }
             //Reading Data
-            LoadData();
+            
+                ResetRecords();
+            
         }
+        private void ResetRecords()
+        {
+            textBox1.Clear();
+            textBox2.Clear();
+            comboBox1.SelectedIndex = -1;
+            button2.Text = "Add";
+            textBox1.Focus();
+        }
+
+        private void button3_Click(object sender, EventArgs e)
+        {
+            ResetRecords();
+        }
+        private bool Validation()
+        {
+            bool result = false;
+            if (!string.IsNullOrEmpty(textBox1.Text) && string.IsNullOrEmpty(textBox2.Text) && comboBox1.SelectedIndex > -1)
+            {
+                result = true;
+            }
+            return result;
+        }
+
     }
 }
